Columns done?
Statistics are saved
Falling blocks keep columns out
No neighbor matches

Not part of the haiku, but I believe this means columns is done, other
than bugs (of which there are probably some that still haven't been
found), dealing with star power, and changing attacks.

I may add some support for the new mechanic we talked about this week
(waiting for an order), but the basic columns game is done.