Some stuff with Multiple Characters
Not much to say here
Lots of bugs sitting in this
Have to commit it

Don't have a ton working really here, just starting to work on multiple
characters for one game, committing so I can use on my laptop when at
office hours.

Only thing that works is that visually you can select your
character....that's it.